improve dsl

+++ Sources/MPSX/ONNX/Nodes/Internal/ReshapeHW.swift
@@ -3,7 +3,7 @@ import MetalPerformanceShadersGraph
 extension MPSGraph {
     func reshapeHW(_ tensor: MPSGraphTensor) -> MPSGraphTensor {
         if let shape = tensor.shape, shape.count == 1 {
-            return tensor.reshape([shape[0], 1, 1])
+            return tensor.reshape([shape[0].intValue, 1, 1])
         }
         return tensor
     }
